# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "my-buddy API",
    "description": "Personal skill-building project companion. Phase 0 scaffold: the API contract is frozen; endpoint bodies are stubs (501) except /health. Contract of record: shared/project-spec.schema.json + this OpenAPI surface (see COORDINATION.md \u00a76).",
    "version": "0.0.0"
  },
  "paths": {
...
    "/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List projects",
        "operationId": "list_projects_projects_get",
        "parameters": [
          {
            "name": "module",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Module"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProjectStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectSummary"
                  },
                  "title": "Response List Projects Projects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Import a spec (validates, persists, runs shop diff \u00a78)",
        "operationId": "create_project_projects_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectSpec"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get hydrated project",
        "operationId": "get_project_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Update project status",
        "operationId": "update_project_projects__project_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Soft-delete a project",
        "operationId": "delete_project_projects__project_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/materials/{material_id}": {
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Toggle a material's checked state",
        "operationId": "update_material_projects__project_id__materials__material_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Material Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/tools/{tool_id}": {
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Toggle a tool's checked / owned state",
        "operationId": "update_tool_projects__project_id__tools__tool_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "tool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tool Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/steps/{step_id}": {
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Toggle a step's completed state / set its note",
        "operationId": "update_step_projects__project_id__steps__step_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "step_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Step Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/retrospective": {
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Upsert the project retrospective",
        "operationId": "upsert_retrospective_projects__project_id__retrospective_patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrospectiveUpsert"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrospectiveRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shop/inventory": {
      "get": {
        "tags": [
          "shop"
        ],
        "summary": "List My Shop",
        "operationId": "list_inventory_shop_inventory_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShopInventoryRead"
                  },
                  "type": "array",
                  "title": "Response List Inventory Shop Inventory Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "shop"
        ],
        "summary": "Add a tool to My Shop",
        "operationId": "add_inventory_shop_inventory_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopInventoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopInventoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shop/inventory/{inventory_id}": {
      "delete": {
        "tags": [
          "shop"
        ],
        "summary": "Remove a tool from My Shop",
        "operationId": "delete_inventory_shop_inventory__inventory_id__delete",
        "parameters": [
          {
            "name": "inventory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Inventory Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaterialRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "est_cost_usd": {
            "type": "number",
            "title": "Est Cost Usd"
          },
          "where_to_find": {
            "type": "string",
            "title": "Where To Find"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "checked": {
            "type": "boolean",
            "title": "Checked",
            "description": "Runtime: toggled from the shopping cart."
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "name",
          "quantity",
          "unit",
          "est_cost_usd",
          "where_to_find",
          "notes",
          "checked"
        ],
        "title": "MaterialRead"
      },
      "MaterialUpdate": {
        "properties": {
          "checked": {
            "type": "boolean",
            "title": "Checked"
          }
        },
        "type": "object",
        "required": [
          "checked"
        ],
        "title": "MaterialUpdate",
        "description": "\u00a711 PATCH /projects/{id}/materials/{mid} \u2014 toggle checked."
      },
...
      "ProjectRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "module_id": {
            "type": "integer",
            "title": "Module Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "Powers URLs, e.g. /my-mechanic/<slug>/\u2026 (\u00a716)."
          },
          "skill_focus": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skill Focus"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "time_budget": {
            "$ref": "#/components/schemas/TimeBudget"
          },
          "estimated_cost_usd": {
            "type": "number",
            "title": "Estimated Cost Usd"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "workspace_required": {
            "type": "string",
            "title": "Workspace Required"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "spec_version": {
            "type": "string",
            "title": "Spec Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/MaterialRead"
            },
            "type": "array",
            "title": "Materials"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/ToolRead"
            },
            "type": "array",
            "title": "Tools"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepRead"
            },
            "type": "array",
            "title": "Steps"
          },
          "research_topics": {
            "items": {
              "$ref": "#/components/schemas/ResearchTopicRead"
            },
            "type": "array",
            "title": "Research Topics"
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/PhotoRead"
            },
            "type": "array",
            "title": "Photos"
          },
          "retrospective": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetrospectiveRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "module_id",
          "name",
          "slug",
          "skill_focus",
          "difficulty",
          "time_budget",
          "estimated_cost_usd",
          "summary",
          "workspace_required",
          "status",
          "spec_version",
          "created_at",
          "materials",
          "tools",
          "steps",
          "research_topics",
          "photos"
        ],
        "title": "ProjectRead",
        "description": "Full hydrated project (\u00a711 GET /projects/{id}): spec + state + research +\nphotos + retrospective."
      },
      "ProjectSpec": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "description": "Spec schema version. Required and pinned; versioned from day one."
          },
          "project": {
            "$ref": "#/components/schemas/ProjectMeta"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/Material"
            },
            "type": "array",
            "title": "Materials"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/Tool"
            },
            "type": "array",
            "title": "Tools"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/Step"
            },
            "type": "array",
            "title": "Steps"
          },
          "research_topics": {
            "items": {
              "$ref": "#/components/schemas/ResearchTopic"
            },
            "type": "array",
            "title": "Research Topics"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "project",
          "materials",
          "tools",
          "steps",
          "research_topics"
        ],
        "title": "ProjectSpec",
        "description": "The full project spec (\u00a76). This is the request body for the import path\n(``POST /projects``) and the response body for generation finalize\n(``POST /generate/sessions/{id}/finalize``)."
      },
      "ProjectStatus": {
        "type": "string",
        "enum": [
          "planning",
          "active",
          "complete"
        ],
        "title": "ProjectStatus",
        "description": "\u00a75 projects.status."
      },
      "ProjectStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ProjectStatusUpdate",
        "description": "\u00a711 PATCH /projects/{id} \u2014 update status."
      },
      "ProjectSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "module_id": {
            "type": "integer",
            "title": "Module Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "Powers URLs, e.g. /my-mechanic/<slug>/\u2026 (\u00a716)."
          },
          "skill_focus": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skill Focus"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "time_budget": {
            "$ref": "#/components/schemas/TimeBudget"
          },
          "estimated_cost_usd": {
            "type": "number",
            "title": "Estimated Cost Usd"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "workspace_required": {
            "type": "string",
            "title": "Workspace Required"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "spec_version": {
            "type": "string",
            "title": "Spec Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "module_id",
          "name",
          "slug",
          "skill_focus",
          "difficulty",
          "time_budget",
          "estimated_cost_usd",
          "summary",
          "workspace_required",
          "status",
          "spec_version",
          "created_at"
        ],
        "title": "ProjectSummary",
        "description": "List view (\u00a711 GET /projects \u2192 the 1a 'Your projects' grid). Row-level\nfields only; no hydrated children."
      },
...
      "RetrospectiveRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "what_went_well": {
            "type": "string",
            "title": "What Went Well"
          },
          "what_id_do_differently": {
            "type": "string",
            "title": "What Id Do Differently"
          },
          "skills_practiced": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Practiced"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "what_went_well",
          "what_id_do_differently",
          "skills_practiced",
          "created_at"
        ],
        "title": "RetrospectiveRead"
      },
      "RetrospectiveUpsert": {
        "properties": {
          "what_went_well": {
            "type": "string",
            "title": "What Went Well"
          },
          "what_id_do_differently": {
            "type": "string",
            "title": "What Id Do Differently"
          },
          "skills_practiced": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Practiced"
          }
        },
        "type": "object",
        "required": [
          "what_went_well",
          "what_id_do_differently",
          "skills_practiced"
        ],
        "title": "RetrospectiveUpsert",
        "description": "\u00a711 PATCH /projects/{id}/retrospective \u2014 upsert retrospective."
      },
...
      "StepRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "order": {
            "type": "integer",
            "title": "Order"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "instruction": {
            "type": "string",
            "title": "Instruction"
          },
          "safety_note": {
            "type": "string",
            "title": "Safety Note"
          },
          "est_time_minutes": {
            "type": "integer",
            "title": "Est Time Minutes"
          },
          "tools_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools Used"
          },
          "materials_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Materials Used"
          },
          "completed": {
            "type": "boolean",
            "title": "Completed",
            "description": "Runtime."
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "Runtime: per-step learning ('where I got stuck')."
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "order",
          "title",
          "instruction",
          "safety_note",
          "est_time_minutes",
          "tools_used",
          "materials_used",
          "completed"
        ],
        "title": "StepRead"
      },
      "StepUpdate": {
        "properties": {
          "completed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "StepUpdate",
        "description": "\u00a711 PATCH /projects/{id}/steps/{sid} \u2014 toggle completed, set note."
      },
...
      "ToolRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "essential": {
            "type": "boolean",
            "title": "Essential"
          },
          "est_cost_usd": {
            "type": "number",
            "title": "Est Cost Usd"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "alternatives": {
            "type": "string",
            "title": "Alternatives"
          },
          "owned": {
            "type": "boolean",
            "title": "Owned",
            "description": "Runtime: set by the shop diff (\u00a78)."
          },
          "acquire": {
            "type": "boolean",
            "title": "Acquire",
            "description": "Runtime: set by the shop diff (\u00a78)."
          },
          "checked": {
            "type": "boolean",
            "title": "Checked",
            "description": "Runtime: for in-cart use."
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "name",
          "essential",
          "est_cost_usd",
          "notes",
          "alternatives",
          "owned",
          "acquire",
          "checked"
        ],
        "title": "ToolRead"
      },
      "ToolUpdate": {
        "properties": {
          "checked": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked"
          },
          "owned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owned"
          }
        },
        "type": "object",
        "title": "ToolUpdate",
        "description": "\u00a711 PATCH /projects/{id}/tools/{tid} \u2014 toggle checked / mark owned."
      },
...
    }
  }
}